+++ extras/video-link-downloader/video-link-downloader.ts
@@ -28,6 +28,12 @@ if (fs.existsSync("settings.json"))
     options = Object.assign(jsonOptions, options);
 }
 
+if (options.parts < 1)
+{
+    console.error("分段数不能小于1");
+    process.exit();
+}
+
 type ProgressHandler = (p: number) => void;
 class Downloader
 {
@@ -105,25 +111,28 @@ class Downloader
         }
         await Promise.all(promises);
         const dest = title + ".flv";
-        return await new Promise<string>(resolve =>
+        console.log("正在合并文件...");
+        if (options.parts === 1)
         {
-            fs.readdir(".", (_, files) =>
+            fs.renameSync(title + ".part0", dest);
+        }
+        else
+        {
+            const files = fs.readdirSync(".");
+            const parts = files.filter(it => it.includes(title + ".part"));
+            const partRegex = /.*\.part([\d]+)/;
+            const data = parts.sort((a, b) =>
             {
-                const parts = files.filter(it => it.includes(title + ".part"));
-                const partRegex = /.*\.part([\d]+)/;
-                const data = parts.sort((a, b) =>
-                {
-                    const partA = parseInt(a.replace(partRegex, "$1"));
-                    const partB = parseInt(b.replace(partRegex, "$1"));
-                    return partA - partB;
-                }).map(file => fs.readFileSync(file));
-                const stream = fs.createWriteStream(dest);
-                data.forEach(it => stream.write(it));
-                stream.close();
-                parts.forEach(file => fs.unlinkSync(file));
-                resolve(dest);
-            });
-        });
+                const partA = parseInt(a.replace(partRegex, "$1"));
+                const partB = parseInt(b.replace(partRegex, "$1"));
+                return partA - partB;
+            }).map(file => fs.readFileSync(file));
+            const stream = fs.createWriteStream(dest);
+            data.forEach(it => stream.write(it));
+            stream.close();
+            parts.forEach(file => fs.unlinkSync(file));
+        }
+        return dest;
     }
     async download()
     {
